Create precedence spec
Holds all the specs related to conflicting configurations and which
takes precedence in which situations
