+++ spec/mimic_spec.rb
@@ -36,6 +36,7 @@ describe 'null object mimicking a class' do
       b.mimic LibraryPatron
     end
   end
+
   it 'responds to all methods defined on the target class' do
     expect(null.member?).to be_nil
     expect(null.name).to be_nil
@@ -52,6 +53,7 @@ describe 'null object mimicking a class' do
     expect(null).to respond_to(:notify_of_overdue_books)
     expect(null).not_to respond_to(:foobar)
   end
+
   it 'has an informative inspect string' do
     expect(null.inspect).to eq('<null:LibraryPatron>')
   end
@@ -78,40 +80,4 @@ describe 'null object mimicking a class' do
       expect(null).to_not respond_to(:login)
     end
   end
-end
-
-describe 'using mimic with black_hole' do
-  subject(:null) { mimic_class.new }
-  let(:mimic_class) do
-    Naught.build do |b|
-      b.mimic Logger
-      b.black_hole
-    end
-  end
-
-  def self.it_behaves_like_a_black_hole_mimic
-    it 'returns self from mimicked methods' do
-      expect(null.info).to equal(null)
-      expect(null.error).to equal(null)
-      expect(null << 'test').to equal(null)
-    end
-
-    it 'does not respond to methods not defined on the target class' do
-      expect { null.foobar }.to raise_error(NoMethodError)
-    end
-  end
-
-  it_behaves_like_a_black_hole_mimic
-
-  describe '(reverse order)' do
-    let(:mimic_class) do
-      Naught.build do |b|
-        b.black_hole
-        b.mimic Logger
-      end
-    end
-
-    it_behaves_like_a_black_hole_mimic
-  end
-
 end
